[Add, Refactor] #6 - model 분리
