+++ 35-seminar/Presentation/Week2/Models/Feedback.swift
@@ -1,0 +1,18 @@
+//
+//  Feedback.swift
+//  35-seminar
+//
+//  Created by 김유림 on 10/26/24.
+//
+
+import Foundation
+
+struct Feedback {
+    var title: String?
+    var author: String?
+    var starCount: Int?
+    var authorDate: Date?
+    var content: String?
+    var developerContent: String?
+    var developerDate: Date?
+}
