+++ 35-seminar/Presentation/Week2/Models/StarColor.swift
@@ -1,0 +1,14 @@
+//
+//  StarColor.swift
+//  35-seminar
+//
+//  Created by 김유림 on 10/26/24.
+//
+
+import Foundation
+
+enum StarColor {
+    case tint
+    case gray
+    case orange
+}
